don't try to sed the log file if it hasn't been created; also move vmargs before the application flag

+++ publish/mirror.sh
@@ -39,6 +39,7 @@ if [[ ${PUBLISH_PATH} != "DO_NOTHING" ]]; then
       -DoutputDirectory=${WORKSPACE}/updates/requirements/ -Dartifact=ant-contrib:ant-contrib:1.0b3:jar
   fi
 
+  mkdir -p ${WORKSPACE}/tmp
   logFile=${WORKSPACE}/tmp/mirror.log.txt
   errFile=${WORKSPACE}/tmp/mirror.err.txt
   rm -f ${logFile} ${errFile}
@@ -46,32 +47,35 @@ if [[ ${PUBLISH_PATH} != "DO_NOTHING" ]]; then
   # get the mirror
   if [[ ${SOURCE_URL} ]]; then SOURCE_URL_PARAM="-DURL=${SOURCE_URL}"; else SOURCE_URL_PARAM=""; fi
   date; time ${JDK8}/bin/java -cp ${WORKSPACE}/eclipse/plugins/org.eclipse.equinox.launcher_*.jar \
-      org.eclipse.equinox.launcher.Main -consoleLog -nosplash -data ${WORKSPACE}/tmp -application org.eclipse.ant.core.antRunner \
-      -f ${SCRIPTNAME} -Dversion=${VERSION} ${SOURCE_URL_PARAM} ${TASK} -vmargs -Declipse.p2.mirrors=false | tee $logFile
+      org.eclipse.equinox.launcher.Main -consoleLog -nosplash -data ${WORKSPACE}/tmp -vmargs -Declipse.p2.mirrors=false \
+      -application org.eclipse.ant.core.antRunner \
+      -f ${SCRIPTNAME} -Dversion=${VERSION} ${SOURCE_URL_PARAM} ${TASK} | tee ${logFile}
 
-  echo "[INFO] Log file: ${logFile}"
+  if [[ -f ${logFile} ]]; then 
+    echo "[INFO] Log file: ${logFile}"
 
-  # check mirror log for failures
-  errorMsgs="java.lang.reflect.InvocationTargetException"
-  errorMsgs="${errorMsgs}|Failed to transfer artifact canonical"
-  errorMsgs="${errorMsgs}|Validation found errors"
-  errorMsgs="${errorMsgs}|Cannot satisfy dependency|Could not resolve content"
-  errorMsgs="${errorMsgs}|Connection refused|Missing requirement"
-  errorMsgs="${errorMsgs}|No repository found|No such file or directory|Unable to read repository"
-  errorMsgs="${errorMsgs}|The following error occurred|BUILD FAIL"
-  sed -n "/${errorMsgs}/p" ${logFile} > ${errFile}
+    # check mirror log for failures
+    errorMsgs="java.lang.reflect.InvocationTargetException"
+    errorMsgs="${errorMsgs}|Failed to transfer artifact canonical"
+    errorMsgs="${errorMsgs}|Validation found errors"
+    errorMsgs="${errorMsgs}|Cannot satisfy dependency|Could not resolve content"
+    errorMsgs="${errorMsgs}|Connection refused|Missing requirement"
+    errorMsgs="${errorMsgs}|No repository found|No such file or directory|Unable to read repository"
+    errorMsgs="${errorMsgs}|The following error occurred|BUILD FAIL"
+    sed -n "/${errorMsgs}/p" ${logFile} > ${errFile}
 
-  if [[ $(cat ${errFile}) ]]; then
-    echo "[ERROR] The following errors have occurred while mirroring - must exit!"
-    echo ""
-    echo "========================================================================"
-    echo ""
-    cat ${errFile}
-    echo ""
-    echo "========================================================================"
-    exit 1
+    if [[ $(cat ${errFile}) ]]; then
+      echo "[ERROR] The following errors have occurred while mirroring - must exit!"
+      echo ""
+      echo "========================================================================"
+      echo ""
+      cat ${errFile}
+      echo ""
+      echo "========================================================================"
+      exit 1
+    fi
   fi
-  
+
   # publish to /builds/staging/${JOB_NAME}_${REQ_NAME}/${VERSION}
   DESTINATION="tools@filemgmt.jboss.org:/downloads_htdocs/tools"
   date
